Pure 2-Days Checkpoint | 20% Overall Completion
This day is exhausting, I was able to make alot of changes especially in Models Part. Models have been too crucial for me because it is been heavily relied by the whole system. I was able to setup most of it, about 100% Implementation and 80% Function Integration. So far, the list below is just an indication of my progress and what I've done in the whole day.

+ Removed Other Apps and Merge Them Into One
+ Created More Variables in Subject_Types.py
- Removed Home_Modals.html
+ Renamed Whole Database
+ Created Multiple Views in One App
+ Created Multiple Instance of URL
+ Created Custom User based on AbstractUser Class
+ Header from | Smart Classroom to | Smart Classroom — System View
+ Renamed Admin/ to Accounts/
+ Added URLs in Sidebar in Every Instance
+ Added AUTH_USER_MODEL in Settings.py
... and so on.

Keep in mind, everything is not working yet! It is best advised to make things right on seperate app instead of making one here. That's all.