Checkpoint | Structure Change
We're able to isolate everything, and used DRY principles in here.
Changes that you may notice here are the following:

+ Change App Name From SC_ControlSystem to SCControlSystem
+ Applied DRY Principle to Render Same Static Required Files with Customizable Renderable Middle Container Content
+ Created Login HTML
+ Restruct Other HTMLs to reside in the folder known as 'template_callable'
+ Removed All Template_View folder in most parts of an app.

it's really hard to make an app that requires relationship instance. Which would lead to forms.py doing explicit actions just to display things right. It might be a wrong solution. Since the app is such compact, I'll be doing it in one app only. No other external apps to be involved in the main project.

We're now attempting to modify views and urls atm, which will be the next to commit.