Deleted Model | SensOutput
Why?
We usually just want to query from the live data. We usually dont have anything to do on it if we got the chance to store that in Database. Seriously, the heck I'm going to do about it? Also the database stirage space will be reduced more as the time progress because we catch that for like 2 times * n requests.